// src/app/types/card.d.ts
export interface CardProps {
  id: number | string;
  src: string;
  alt: string;
  title: string;
  category: string;
  description: string;
  price: number;
  favorite: boolean;
}

// ,
// {
//   "id": 2,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }
// ,
// {
//   "id": 3,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// },
// {
//   "id": 4,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }
// ,
// {
//   "id": 5,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }
// ,
// {
//   "id": 6,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }
// ,
// {
//   "id": 7,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }
// ,
// {
//   "id": 8,
//   "src": "/path/to/image2.jpg",
//   "alt": "Image 2",
//   "title": "Card 2",
//   "description": "This is the second card.",
//   "price": 49.99,
//   "favorite": true
// }